requirements.txt
ebooklib>=0.17.1
pypandoc>=1.11
Pillow>=10.0.0
flask>=2.0.1
werkzeug>=2.0.1
gunicorn>=20.1.0
